requirements.txt
# Basis-Tools
wheel>=0.42.0
setuptools>=69.0.0

# Web Framework
flask==3.0.2
flask-restx==1.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Konfiguration
pyyaml==6.0.2
python-dotenv>=0.19.0

# MongoDB
pymongo==4.13.2
dnspython==2.6.1

# Testing
pytest==8.1.1
pytest-asyncio>=0.23.5

# Bildverarbeitung
pillow==11.1.0 --only-binary :all:
pytesseract==0.3.10
pdf2image==1.17.0

# Dokumentenverarbeitung
PyMuPDF>=1.25.0
pypdf>=4.1.0
# Office (python-only Pipeline A)
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==0.6.23
#python-magic-bin==0.4.14 #This is only available on windows, so it wont work in a linux docker container
python-magic>=0.4.27



# Audio/Video Verarbeitung
numpy>=1.24.0

# Datenverarbeitung
pandas>=1.0.0
psutil>=5.9.0

# OpenAI Integration
openai>=1.12.0
tiktoken>=0.6.0

# Mistral AI Integration
mistralai>=1.9.0

# YouTube Download
yt-dlp==2024.3.10

# Development
mypy==1.8.0

# No idea does not work without it
pydub>=0.25.1

markdown

# RAG / Vector Search
voyageai>=0.2.0

# JSON Schema Validation
jsonschema>=4.20.0